Make Printer:::table_2 more generic
This change makes table_2 be able to accept more than just a Vec<(String, String)> which is what it was constrained to before.

The function can now take in anything that implements IntoIterator (https://doc.rust-lang.org/std/iter/trait.IntoIterator.html#implementors) including iterators themselves and, as before, Vec. Th efunction is also now generic over the type of the elements in the tuple, allowing for anything that implements AsRef<str> (String, &str, Cow<str>, ...) this functionality is the same of that in the `Printer::println` function and makes writing code much easier since callers do not need to deal with tons of `.to_string()`s on their string literals